+++ src/helper.ts
@@ -5,10 +5,11 @@ import { RequestInit } from "node-fetch";
 import RestApi from '@signageos/sdk/dist/RestApi/RestApi';
 import IRestApiOptions from '@signageos/sdk/dist/RestApi/IOptions';
 import IRestApiAccountOptions from '@signageos/sdk/dist/RestApi/IOptions';
-import { IConfig, loadConfig } from './RunControl/runControlHelper';
+import { loadConfig } from './RunControl/runControlHelper';
 import { ApiVersions } from '@signageos/sdk/dist/RestApi/apiVersions';
 import { parameters } from './parameters';
 import { getGlobalApiUrl } from './Command/globalArgs';
+import { IConfig } from '@signageos/sdk/dist/SosHelper/sosControlHelper';
 
 interface ICredentials {
 	oauthClientId: string;