+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Console output of `npm start` of generated applet shows correct URL of emulator
+- The `--entry-file` can be now ommited when the `package.json` has the `main` property set
+- The `sos device connect` is looking for the applet in server by `name` instead of required `uid` in `package.json` `sos.appletUid` property
+- The `.git` folder is ignored automatically when the `.gitignore` file is used
+
+### Added
+- Allow customize `--server-port` and `--server-public-url` for `sos device connect` command
+
+### Deprecated
+- Remove support of experimental version of webpack-plugin v0.2. Use version v1+ instead
 
 ## [1.1.5] - 2023-01-02
 ### Fixed